Create Automating Packaged Components and Deployments
This process demonstrates how to use the AtomSphere API to create a packaged component version for a process and deploy it to an environment. After the PackagedComponent is created, the newly created ID is stored in a Process Property for use in the create DeployedPackage call in the second branch. Note that in this example the details for the process to package and environment to which to deploy are defined in static message shapes but in practice these values might be passed dynamically from an external client. Also, in practice it might be preferred to implement these two actions as separate processes as part of an automated CI/CD workflow.